refactor: move run to lib/infra/
Move the run engine module (24 source files + 4 test files) from
src/lib/run/ to src/lib/infra/run/ as part of the lib/ restructuring
effort (#7697). This is the largest infra module with DB schema
dependencies and heavy cross-module imports.

Closes #7726
