+++ turbo/apps/web/app/api/agent/runs/[id]/cancel/route.ts
@@ -13,7 +13,7 @@ import { eq, and } from "drizzle-orm";
 import {
   cancelRun,
   dispatchCancelSideEffects,
-} from "../../../../../../src/lib/run/run-service";
+} from "../../../../../../src/lib/infra/run/run-service";
 import {
   dispatchQueuedZeroRun,
   drainOrgQueue,
