+++ turbo/apps/web/app/api/agent/runs/[id]/route.ts
@@ -6,7 +6,7 @@ import {
   isAuthError,
 } from "../../../../../src/lib/auth/require-auth";
 import { resolveOrg } from "../../../../../src/lib/org/resolve-org";
-import { getRunById } from "../../../../../src/lib/run/run-service";
+import { getRunById } from "../../../../../src/lib/infra/run/run-service";
 
 const router = tsr.router(runsByIdContract, {
   getById: async ({ params, headers }) => {
