+++ turbo/apps/web/app/api/agent/runs/[id]/telemetry/agent/route.ts
@@ -7,7 +7,7 @@ import { runAgentEventsContract } from "@vm0/core";
 import { initServices } from "../../../../../../../src/lib/init-services";
 import { getAuthContext } from "../../../../../../../src/lib/auth/get-auth-context";
 import { resolveOrg } from "../../../../../../../src/lib/org/resolve-org";
-import { getRunAgentEvents } from "../../../../../../../src/lib/run/run-telemetry-service";
+import { getRunAgentEvents } from "../../../../../../../src/lib/infra/run/run-telemetry-service";
 import { isNotFound } from "../../../../../../../src/lib/errors";
 
 const router = tsr.router(runAgentEventsContract, {
