+++ turbo/apps/web/app/api/internal/callbacks/chat/route.ts
@@ -6,7 +6,7 @@ import { verifyCallback } from "../../../../../src/lib/callback";
 import { chatThreads } from "../../../../../src/db/schema/chat-thread";
 import { agentRuns } from "../../../../../src/db/schema/agent-run";
 import { findNewSessionId } from "../../../../../src/lib/session/find-new-session";
-import { queryRunEventsForChat } from "../../../../../src/lib/run/extract-chat-events";
+import { queryRunEventsForChat } from "../../../../../src/lib/infra/run/extract-chat-events";
 import { appendChatMessages } from "../../../../../src/lib/zero/zero-session-service";
 import {
   generateChatTitle,
