+++ turbo/apps/web/app/api/zero/email/callbacks/trigger/route.ts
@@ -4,7 +4,7 @@ import { initServices } from "../../../../../../src/lib/init-services";
 import { verifyCallback } from "../../../../../../src/lib/callback";
 import { agentRuns } from "../../../../../../src/db/schema/agent-run";
 import { zeroAgents } from "../../../../../../src/db/schema/zero-agent";
-import { getRunOutputText } from "../../../../../../src/lib/run/extract-run-output";
+import { getRunOutputText } from "../../../../../../src/lib/infra/run/extract-run-output";
 import { enqueueEmail } from "../../../../../../src/lib/email/outbox-service";
 import {
   buildReplyToAddress,
