+++ turbo/apps/web/app/api/zero/runs/[id]/context/route.ts
@@ -10,8 +10,8 @@ import {
   isAuthError,
 } from "../../../../../../src/lib/auth/require-auth";
 import { resolveOrg } from "../../../../../../src/lib/org/resolve-org";
-import { getRunById } from "../../../../../../src/lib/run/run-service";
-import { queryRunContext } from "../../../../../../src/lib/run/run-context-service";
+import { getRunById } from "../../../../../../src/lib/infra/run/run-service";
+import { queryRunContext } from "../../../../../../src/lib/infra/run/run-context-service";
 
 const router = tsr.router(zeroRunContextContract, {
   getContext: async ({ params, headers }) => {
