+++ turbo/apps/web/app/api/zero/runs/[id]/telemetry/agent/route.ts
@@ -10,7 +10,7 @@ import {
   isAuthError,
 } from "../../../../../../../src/lib/auth/require-auth";
 import { resolveOrg } from "../../../../../../../src/lib/org/resolve-org";
-import { getRunAgentEvents } from "../../../../../../../src/lib/run/run-telemetry-service";
+import { getRunAgentEvents } from "../../../../../../../src/lib/infra/run/run-telemetry-service";
 import { isNotFound } from "../../../../../../../src/lib/errors";
 
 const router = tsr.router(zeroRunAgentEventsContract, {
