+++ turbo/apps/web/src/lib/infra/run/__tests__/extract-run-output.test.ts
@@ -11,7 +11,7 @@ vi.mock("@axiomhq/js", () => {
   };
 });
 
-import { reloadEnv } from "../../../env";
+import { reloadEnv } from "../../../../env";
 import {
   extractRunOutput,
   extractAllRunOutputs,
