+++ turbo/apps/web/src/lib/infra/run/__tests__/run-status.test.ts
@@ -3,12 +3,12 @@ import {
   testContext,
   uniqueId,
   type UserContext,
-} from "../../../__tests__/test-helpers";
+} from "../../../../__tests__/test-helpers";
 import {
   createTestCompose,
   createTestRunInDb,
   findTestRunRecord,
-} from "../../../__tests__/api-test-helpers";
+} from "../../../../__tests__/api-test-helpers";
 import { transitionRunStatus } from "../run-status";
 
 const context = testContext();
