+++ turbo/apps/web/src/lib/infra/run/context/build-context.ts
@@ -1,6 +1,6 @@
 import { expandEnvironmentFromCompose } from "../environment/expand-environment";
 import type { ExecutionContext, ResumeSession } from "../types";
-import type { ArtifactSnapshot } from "../../checkpoint/types";
+import type { ArtifactSnapshot } from "../../../checkpoint/types";
 import type { Firewalls } from "@vm0/core";
 
 interface BuildInfraContextParams {
