+++ turbo/apps/web/src/lib/infra/run/extract-run-output.ts
@@ -1,4 +1,4 @@
-import { queryAxiom, getDatasetName, DATASETS } from "../shared/axiom";
+import { queryAxiom, getDatasetName, DATASETS } from "../../shared/axiom";
 
 // ---------------------------------------------------------------------------
 // Types
