+++ turbo/apps/web/src/lib/infra/run/resolvers/resolve-session.ts
@@ -2,11 +2,11 @@ import { eq } from "drizzle-orm";
 import {
   agentComposes,
   agentComposeVersions,
-} from "../../../db/schema/agent-compose";
-import { agentRuns } from "../../../db/schema/agent-run";
-import { notFound, unauthorized, badRequest } from "../../errors";
-import { logger } from "../../logger";
-import { getAgentSessionWithConversation } from "../../agent-session";
+} from "../../../../db/schema/agent-compose";
+import { agentRuns } from "../../../../db/schema/agent-run";
+import { notFound, unauthorized, badRequest } from "../../../errors";
+import { logger } from "../../../logger";
+import { getAgentSessionWithConversation } from "../../../agent-session";
 import type { ConversationResolution } from "./types";
 import { extractWorkingDir, extractCliAgentType } from "../utils";
 import { resolveSessionHistory } from "./resolve-session-history";
