+++ turbo/apps/web/src/lib/infra/run/run-status.ts
@@ -1,9 +1,9 @@
 import { and, eq, inArray } from "drizzle-orm";
-import { agentRuns } from "../../db/schema/agent-run";
-import { dispatchCallbacks } from "../callback";
-import { logger } from "../logger";
+import { agentRuns } from "../../../db/schema/agent-run";
+import { dispatchCallbacks } from "../../callback";
+import { logger } from "../../logger";
 import type { RunResult, RunStatus } from "./types";
-import type { Database } from "../../types/global";
+import type { Database } from "../../../types/global";
 
 const log = logger("service:run-status");
