+++ turbo/apps/web/src/lib/infra/run/run-telemetry-service.ts
@@ -1,8 +1,8 @@
 import { eq, and } from "drizzle-orm";
-import { agentRuns } from "../../db/schema/agent-run";
-import { agentComposeVersions } from "../../db/schema/agent-compose";
-import { queryAxiom, getDatasetName, DATASETS } from "../shared/axiom";
-import { notFound } from "../errors";
+import { agentRuns } from "../../../db/schema/agent-run";
+import { agentComposeVersions } from "../../../db/schema/agent-compose";
+import { queryAxiom, getDatasetName, DATASETS } from "../../shared/axiom";
+import { notFound } from "../../errors";
 
 interface AxiomAgentEvent {
   _time: string;
